Fix database structure for better normalization
